requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
alembic
psycopg2-binary
